# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SendOtpRequest {
  /** Phone number in E.164 format */
  phone: string;
}

export interface SendOtpResponse {
  message: string;
  /** OTP shown only in development mode */
  devOtp?: string;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
}

export interface User {
  id: number;
  phone: string;
  name?: string | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type MeasurementInputFitPreference =
  (typeof MeasurementInputFitPreference)[keyof typeof MeasurementInputFitPreference];

export const MeasurementInputFitPreference = {
  slim: "slim",
  regular: "regular",
  relaxed: "relaxed",
} as const;

export interface MeasurementInput {
  waist: number;
  hip: number;
  pantLength: number;
  thigh: number;
  fitPreference: MeasurementInputFitPreference;
}

export interface Measurement {
  id: number;
  userId: number;
  waist: number;
  hip: number;
  pantLength: number;
  thigh: number;
  fitPreference: string;
  updatedAt: string;
}

export type OrderInputFabricType =
  (typeof OrderInputFabricType)[keyof typeof OrderInputFabricType];

export const OrderInputFabricType = {
  cotton: "cotton",
  "stretch-cotton": "stretch-cotton",
  linen: "linen",
} as const;

export type OrderInputPocketStyle =
  (typeof OrderInputPocketStyle)[keyof typeof OrderInputPocketStyle];

export const OrderInputPocketStyle = {
  classic: "classic",
  cargo: "cargo",
  "no-pocket": "no-pocket",
} as const;

export type OrderInputOccasion =
  (typeof OrderInputOccasion)[keyof typeof OrderInputOccasion];

export const OrderInputOccasion = {
  formal: "formal",
  casual: "casual",
} as const;

export interface OrderInput {
  fabricType: OrderInputFabricType;
  color: string;
  pocketStyle: OrderInputPocketStyle;
  occasion: OrderInputOccasion;
  measurementId: number;
}

export interface Order {
  id: number;
  userId: number;
  measurementId: number;
  fabricType: string;
  color: string;
  pocketStyle: string;
  occasion: string;
  estimatedPrice: number;
  status: string;
  createdAt: string;
  measurement?: Measurement;
}

export type AdminOrder = Order & {
  user?: User;
};

export type AdminMeasurement = Measurement & {
  user?: User;
};