# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send OTP to phone number
 */
export const SendOtpBody = zod.object({
  phone: zod.string().describe("Phone number in E.164 format"),
});

export const SendOtpResponse = zod.object({
  message: zod.string(),
  devOtp: zod
    .string()
    .optional()
    .describe("OTP shown only in development mode"),
});

/**
 * @summary Verify OTP and login/signup
 */
export const VerifyOtpBody = zod.object({
  phone: zod.string(),
  otp: zod.string(),
});

export const VerifyOtpResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    phone: zod.string(),
    name: zod.string().nullish(),
    isAdmin: zod.boolean(),
    createdAt: zod.date(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  phone: zod.string(),
  name: zod.string().nullish(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Logout user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Save user measurements
 */
export const SaveMeasurementsBody = zod.object({
  waist: zod.number(),
  hip: zod.number(),
  pantLength: zod.number(),
  thigh: zod.number(),
  fitPreference: zod.enum(["slim", "regular", "relaxed"]),
});

export const SaveMeasurementsResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  waist: zod.number(),
  hip: zod.number(),
  pantLength: zod.number(),
  thigh: zod.number(),
  fitPreference: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary Get user measurements
 */
export const GetMeasurementsResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  waist: zod.number(),
  hip: zod.number(),
  pantLength: zod.number(),
  thigh: zod.number(),
  fitPreference: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  fabricType: zod.enum(["cotton", "stretch-cotton", "linen"]),
  color: zod.string(),
  pocketStyle: zod.enum(["classic", "cargo", "no-pocket"]),
  occasion: zod.enum(["formal", "casual"]),
  measurementId: zod.number(),
});

/**
 * @summary Get user orders
 */
export const GetOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  measurementId: zod.number(),
  fabricType: zod.string(),
  color: zod.string(),
  pocketStyle: zod.string(),
  occasion: zod.string(),
  estimatedPrice: zod.number(),
  status: zod.string(),
  createdAt: zod.date(),
  measurement: zod
    .object({
      id: zod.number(),
      userId: zod.number(),
      waist: zod.number(),
      hip: zod.number(),
      pantLength: zod.number(),
      thigh: zod.number(),
      fitPreference: zod.string(),
      updatedAt: zod.date(),
    })
    .optional(),
});
export const GetOrdersResponse = zod.array(GetOrdersResponseItem);

/**
 * @summary Get a single order
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  measurementId: zod.number(),
  fabricType: zod.string(),
  color: zod.string(),
  pocketStyle: zod.string(),
  occasion: zod.string(),
  estimatedPrice: zod.number(),
  status: zod.string(),
  createdAt: zod.date(),
  measurement: zod
    .object({
      id: zod.number(),
      userId: zod.number(),
      waist: zod.number(),
      hip: zod.number(),
      pantLength: zod.number(),
      thigh: zod.number(),
      fitPreference: zod.string(),
      updatedAt: zod.date(),
    })
    .optional(),
});

/**
 * @summary Admin - Get all users
 */
export const AdminGetUsersResponseItem = zod.object({
  id: zod.number(),
  phone: zod.string(),
  name: zod.string().nullish(),
  isAdmin: zod.boolean(),
  createdAt: zod.date(),
});
export const AdminGetUsersResponse = zod.array(AdminGetUsersResponseItem);

/**
 * @summary Admin - Get all orders
 */
export const AdminGetOrdersResponseItem = zod
  .object({
    id: zod.number(),
    userId: zod.number(),
    measurementId: zod.number(),
    fabricType: zod.string(),
    color: zod.string(),
    pocketStyle: zod.string(),
    occasion: zod.string(),
    estimatedPrice: zod.number(),
    status: zod.string(),
    createdAt: zod.date(),
    measurement: zod
      .object({
        id: zod.number(),
        userId: zod.number(),
        waist: zod.number(),
        hip: zod.number(),
        pantLength: zod.number(),
        thigh: zod.number(),
        fitPreference: zod.string(),
        updatedAt: zod.date(),
      })
      .optional(),
  })
  .and(
    zod.object({
      user: zod
        .object({
          id: zod.number(),
          phone: zod.string(),
          name: zod.string().nullish(),
          isAdmin: zod.boolean(),
          createdAt: zod.date(),
        })
        .optional(),
    }),
  );
export const AdminGetOrdersResponse = zod.array(AdminGetOrdersResponseItem);

/**
 * @summary Admin - Get all measurements
 */
export const AdminGetMeasurementsResponseItem = zod
  .object({
    id: zod.number(),
    userId: zod.number(),
    waist: zod.number(),
    hip: zod.number(),
    pantLength: zod.number(),
    thigh: zod.number(),
    fitPreference: zod.string(),
    updatedAt: zod.date(),
  })
  .and(
    zod.object({
      user: zod
        .object({
          id: zod.number(),
          phone: zod.string(),
          name: zod.string().nullish(),
          isAdmin: zod.boolean(),
          createdAt: zod.date(),
        })
        .optional(),
    }),
  );
export const AdminGetMeasurementsResponse = zod.array(
  AdminGetMeasurementsResponseItem,
);