# lib/api-zod/src/generated/types/adminMeasurement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
import type { Measurement } from "./measurement";
import type { User } from "./user";

export type AdminMeasurement = Measurement & {
  user?: User;
};