# lib/api-zod/src/generated/types/measurement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */

export interface Measurement {
  id: number;
  userId: number;
  waist: number;
  hip: number;
  pantLength: number;
  thigh: number;
  fitPreference: string;
  updatedAt: Date;
}