# lib/api-zod/src/generated/types/measurementInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementInputFitPreference } from "./measurementInputFitPreference";

export interface MeasurementInput {
  waist: number;
  hip: number;
  pantLength: number;
  thigh: number;
  fitPreference: MeasurementInputFitPreference;
}