# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
import type { Measurement } from "./measurement";

export interface Order {
  id: number;
  userId: number;
  measurementId: number;
  fabricType: string;
  color: string;
  pocketStyle: string;
  occasion: string;
  estimatedPrice: number;
  status: string;
  createdAt: Date;
  measurement?: Measurement;
}