# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputFabricType } from "./orderInputFabricType";
import type { OrderInputOccasion } from "./orderInputOccasion";
import type { OrderInputPocketStyle } from "./orderInputPocketStyle";

export interface OrderInput {
  fabricType: OrderInputFabricType;
  color: string;
  pocketStyle: OrderInputPocketStyle;
  occasion: OrderInputOccasion;
  measurementId: number;
}