# lib/api-zod/src/generated/types/sendOtpRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CustomFit Pants API
 * OpenAPI spec version: 0.1.0
 */

export interface SendOtpRequest {
  /** Phone number in E.164 format */
  phone: string;
}